Move backend_opt fixture
